[IMP]hotel: improved the code and fixed issue
