+++ hotel/views/hotel_folio.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8" ?>
 <odoo>
     <!-- Form view of hotel folio -->
     <record id="view_hotel_folio_form" model="ir.ui.view">
@@ -26,22 +25,19 @@
                         states="cancel"
                         string="Set to Draft"
                         type="object"
-                        icon="fa-check-square-o"
                         class="oe_highlight"
                     />
                     <button
                         name="action_cancel"
-                        string="Cancel Folio"
+                        string="Cancel"
                         states="sale"
                         type="object"
-                        icon="fa-close"
                         class="oe_highlight"
                     />
                     <button
                         name="action_cancel"
-                        string="Cancel Folio"
+                        string="Cancel"
                         states="draft"
-                        icon="fa-close"
                         type="object"
                         class="oe_highlight"
                     />
@@ -107,7 +103,8 @@
                                 colspan="4"
                                 string="Room Line"
                                 nolabel="1"
-                                context="{'default_checkin_date':'checkin_date','default_checkout_date':'checkout_date'}"
+                                context="{'default_checkin_date':checkin_date,
+                                        'default_checkout_date':checkout_date}"
                             >
                                 <tree editable="bottom" string="Room Line">
                                     <field
@@ -171,8 +168,9 @@
                                 colspan="4"
                                 string="Service Line"
                                 nolabel="1"
-                                context="{'deafult_ser_checkin_date':'checkin_date',
-                                'deafult_ser_checkout_date':'checkout_date','checkin':checkin_date,'checkout':checkout_date}"
+                                context="{'deafult_ser_checkin_date':checkin_date,
+                                'deafult_ser_checkout_date':checkout_date
+                                }"
                             >
 
                                 <tree editable="bottom" string="Service Line">
